[RFC] Fix multiple issues with signing and testing
Exploratory solution. Changes in this commit should be split into
multiple, separate commits.

* The epub and wikitext core jars were not signed by the Eclipse
jarsigner web service.

* Ensure that the jarsigner execution runs before the GPG plugin
execution in accordance with
https://wiki.eclipse.org/IT_Infrastructure_Doc#What_about_GPG_signing.3F

* The epub UI plugin manifest also specified a very old version of the
org.eclipse.mylyn.wikitext as a require-bundle dependency.

* Add the tycho gpg plugin to ensure the manifest-first plugins are
GPG signed prior to deploying the p2 repository on the
download.eclipse.org fileshare.
-- TODO: configuration of this plugin needs adjustment.

* The org.eclipse.mylyn.wiktext.tests project has integration test
against wikitext. But the test classes are placed in java packages
that also exist in the wikitext plugins that it's testing. This
causes problems when jar signing is enabled (-Psign) because the
dependent wikitext plugins will be signed but the integration test
classes in org.eclipse.mylyn.wiktext.tests have not by the time that
surefire is running. This results in a message like the following:

'class "...DocBookDocumentBuilderIntegrationTest"'s signer information
does not match signer information of other classes in the same package'

After exploring and failing to find a way to configure surefire or
failsafe to workaround this scenario, the integration test classes in
org.eclipse.mylyn.wikitext.tests have been moved to a separate java
package that does not collide with packages in any wikitext plugin.
There's no reason the integration test classes needed to be in the same
package. However, this 'solution' feels wrong :( and I'm open to
other ideas for resolving this.