Fix regression in wikitext maven plugin when copying non-markup files
The wikitext maven plugin exists to facilitate creating HTML
documentation from eclipse help source files. A regression was
introduced with commit
803292b246c65c115333d12b2730a3b9759bb397 that switched out the Guava
Files.copy method for the java.nio.file.Files.copy method.

The semantics of those 2 methods are different. While Guava's copy
method will always replace/overwrite the target file if it already
exists, the java.nio.file.Files.copy method does not do that by
default. The optional StandardCopyOption.REPLACE_EXISTING must
be provided as a parameter to allow for the same behavior. This
change is important if the workspace is not clean when the
wikitext.maven mojo is executed.